+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `logOutFromSession` mutation
+
 ## [2.140.0] - 2021-04-14
+
 ### Added
 - `loginSessionsInfo` query